Moving admin user creation to the Quarkus startup phase
Closes #29072
